requirements.txt
amqp==2.6.1
asgiref==3.5.0
billiard==3.6.4.0
celery==4.4.2
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
Django==3.2.11
django-rest-knox==4.1.0
djangorestframework==3.12.4
drf-nested-routers==0.93.4
flake8==4.0.1
kombu==4.6.11
mccabe==0.6.1
prompt-toolkit==3.0.26
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
vine==1.3.0
wcwidth==0.2.5
